Enhancements to User APIs
- Added @Transactional annotation to createUser() method
  so that any requests sent by user are saved only if there are
  no errors.
- Changed user_id to be auto generated by the database, and used
  PreparedSytatement to insert into tables.
- Added SQL files.

+++ src/main/java/com/TimeNexus/TimeNexus/dto/User.java
@@ -11,7 +11,7 @@ import org.springframework.data.annotation.Id;
 public class User {
 
     @Id
-    private Long user_id;
+    private int user_id;
     private String first_name;
     private String last_name;
     private String email;
